+++ src/aco/probe/csv_probe.rs
@@ -1,6 +1,6 @@
 use serde::Serialize;
 
-use crate::aco::ants_system::probe::Probe;
+use crate::aco::probe::Probe;
 use crate::aco::FMatrix;
 use crate::aco::Solution;
